refactor: move audio time adj fns to audio_utils.py
align.py is getting too big, and I'm going to add more with my work on anchors.
Since calculate_adjustment and correct_adjustment are about time stamps in audio
files, even though they are not manipulating the audio streams themselves, I
decided they fit better in audio_utils.py that in the oversized align.py.
