+++ readalongs/align.py
@@ -9,7 +9,6 @@
 #
 #######################################################################
 
-import copy
 import io
 import os
 import shutil
@@ -24,7 +23,14 @@ from pydub.exceptions import CouldntEncodeError
 from pympi.Praat import TextGrid
 from webvtt import Caption, WebVTT
 
-from readalongs.audio_utils import mute_section, read_audio_from_file, remove_section
+from readalongs.audio_utils import (
+    calculate_adjustment,
+    correct_adjustments,
+    mute_section,
+    read_audio_from_file,
+    remove_section,
+    sort_and_join_dna_segments,
+)
 from readalongs.log import LOGGER
 from readalongs.portable_tempfile import PortableNamedTemporaryFile
 from readalongs.text.add_elements_to_xml import add_images, add_supplementary_xml
@@ -35,53 +41,6 @@ from readalongs.text.make_fsg import make_fsg
 from readalongs.text.make_smil import make_smil
 from readalongs.text.tokenize_xml import tokenize_xml
 from readalongs.text.util import save_minimal_index_html, save_txt, save_xml
-
-
-def sort_and_join_dna_segments(do_not_align_segments: List[dict]) -> List[dict]:
-    """ Give a list of DNA segments, sort them and join any overlapping ones """
-    results = []
-    for seg in sorted(do_not_align_segments, key=lambda x: x["begin"]):
-        if results and results[-1]["end"] >= seg["begin"]:
-            results[-1]["end"] = max(results[-1]["end"], seg["end"])
-        else:
-            results.append(copy.deepcopy(seg))
-    return results
-
-
-def correct_adjustments(
-    start: int, end: int, do_not_align_segments: List[dict]
-) -> Tuple[int, int]:
-    """ Given the start and end of a segment (in ms) and a list of do-not-align segments,
-        If one of the do-not-align segments occurs inside one of the start-end range,
-        align the start or end with the do-not-align segment, whichever requires minimal change
-    """
-    for seg in do_not_align_segments:
-        if start < seg["begin"] and end > seg["end"]:
-            if seg["begin"] - start > end - seg["end"]:
-                return start, seg["begin"]
-            else:
-                return seg["end"], end
-    return start, end
-
-
-def calculate_adjustment(timestamp: int, do_not_align_segments: List[dict]) -> int:
-    """ Given a time (in ms) and a list of do-not-align segments,
-        return the sum (ms) of the lengths of the do-not-align segments
-        that start before the timestamp
-
-    Preconditions:
-        do_not_align_segments are sorted in ascending order of their "begin" and do not overlap
-    """
-    results = 0
-    prev_end = -1
-    for seg in do_not_align_segments:
-        assert prev_end < seg["begin"]
-        prev_end = seg["end"]
-        if seg["begin"] <= timestamp:
-            delta = seg["end"] - seg["begin"]
-            results += delta
-            timestamp += delta
-    return results
 
 
 def align_audio(  # noqa: C901
