+++ test/test_audio.py
@@ -7,16 +7,14 @@ from shutil import rmtree
 from subprocess import run
 from unittest import TestCase, main
 
-from readalongs.align import (
+from readalongs.audio_utils import (
     calculate_adjustment,
     correct_adjustments,
-    sort_and_join_dna_segments,
-)
-from readalongs.audio_utils import (
     join_section,
     mute_section,
     read_audio_from_file,
     remove_section,
+    sort_and_join_dna_segments,
     write_audio_to_file,
 )
 from readalongs.log import LOGGER
